requirements.txt
# Core
torch>=2.0
torchvision>=0.15
spconv-cu120  # or spconv-cu118 for CUDA 11.8

# Data & computation
numpy>=1.24
scipy>=1.10
opencv-python-headless>=4.8
h5py>=3.8
PyYAML>=6.0

# Training utilities
tqdm>=4.65
matplotlib>=3.7
pandas>=2.0
scikit-learn>=1.3

# SNN backbone
spikingjelly==0.0.0.0.14
